Add time_record object.
